+++ designer/client/src/reducers/graph/utils.ts
@@ -124,8 +124,8 @@ export function enrichNodeWithProcessDependentData(
 
     switch (NodeUtils.nodeType(node)) {
         case "Join": {
-            const { parameters } = ProcessUtils.extractComponentDefinition(node, processDefinitionData.components);
-            const declaredBranchParameters = parameters.filter((p) => p.branchParam);
+            const parameters = ProcessUtils.extractComponentDefinition(node, processDefinitionData.components)?.parameters;
+            const declaredBranchParameters = parameters?.filter((p) => p.branchParam) || [];
             const incomingEdges = edges.filter((e) => e.to === node.id);
             const branchParameters = incomingEdges.map((edge) => {
                 const branchId = edge.from;