+++ designer/client/src/types/component.ts
@@ -1,10 +1,5 @@
 import { BranchParametersTemplate, NodeType } from "./node";
 
-export type SingleComponentConfig = {
-    icon?: string;
-    docsUrl?: string;
-    componentGroup?: string;
-};
 export type Component = {
     branchParametersTemplate: BranchParametersTemplate;
     node: NodeType;