+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/definition/DefinitionsService.scala
@@ -6,8 +6,6 @@ import pl.touk.nussknacker.engine.api.deployment.DeploymentManager
 import pl.touk.nussknacker.engine.api.process.ProcessingType
 import pl.touk.nussknacker.engine.definition.component.{ComponentStaticDefinition, FragmentSpecificData}
 import pl.touk.nussknacker.engine.definition.model.ModelDefinition
-import pl.touk.nussknacker.engine.graph.expression.Expression
-import pl.touk.nussknacker.engine.modelconfig.ComponentsUiConfigParser
 import pl.touk.nussknacker.engine.util.Implicits.RichScalaMap
 import pl.touk.nussknacker.engine.{ModelData, ProcessingTypeData}
 import pl.touk.nussknacker.restmodel.definition._
@@ -58,7 +56,6 @@ class DefinitionsService(
       components =
         modelDefinitionWithBuiltInComponentsAndFragments.components.mapValuesNow(createUIComponentDefinition),
       classes = modelData.modelDefinitionWithClasses.classDefinitions.all.toList.map(_.clazzName),
-      componentsConfig = prepareComponentConfig(modelDefinitionWithBuiltInComponentsAndFragments, modelData),
       scenarioPropertiesConfig =
         (if (forFragment) FragmentPropertiesConfig.properties else finalizedScenarioPropertiesConfig)
           .mapValuesNow(createUIScenarioPropertyConfig),
@@ -68,32 +65,14 @@ class DefinitionsService(
     )
   }
 
-  private def prepareComponentConfig(
-      modelDefinitionWithBuiltInComponentsAndFragments: ModelDefinition[ComponentStaticDefinition],
-      modelData: ModelData
-  ) = {
-    modelDefinitionWithBuiltInComponentsAndFragments.components.map { case (info, value) =>
-      info.name -> value.componentConfig
-    } ++ preparePropertiesConfig(modelData)
-  }
-
-  // TODO - Extract to the separate, named field in UIDefinitions which would hold also scenarioPropertiesConfig
-  //      - Stop treating properties as a node on FE side
-  //      - Other way to configure it - it should be somewhere around scenarioPropertiesConfig
-  //      - Documentation
-  // TODO (alternative): get rid of support fot that, we can configure only icon and docsUrl thanks to that?
-  private def preparePropertiesConfig(modelData: ModelData) = {
-    val componentsUiConfig          = ComponentsUiConfigParser.parse(modelData.modelConfig)
-    val propertiesFakeComponentName = "$properties"
-    componentsUiConfig.componentsConfig.get(propertiesFakeComponentName).map(propertiesFakeComponentName -> _)
-  }
-
   private def createUIComponentDefinition(
       componentDefinition: ComponentStaticDefinition
   ): UIComponentDefinition = {
     UIComponentDefinition(
       parameters = componentDefinition.parameters.map(createUIParameter),
       returnType = componentDefinition.returnType,
+      icon = componentDefinition.iconUnsafe,
+      docsUrl = componentDefinition.componentConfig.docsUrl,
       outputParameters = Option(componentDefinition.componentTypeSpecificData).collect {
         case FragmentSpecificData(outputNames) => outputNames
       }
@@ -120,12 +99,11 @@ object DefinitionsService {
     )
 
   def createUIParameter(parameter: Parameter): UIParameter = {
-    val defaultValue = parameter.defaultValue.getOrElse(Expression.spel(""))
     UIParameter(
       name = parameter.name,
       typ = parameter.typ,
-      editor = parameter.editor.getOrElse(RawParameterEditor),
-      defaultValue = defaultValue,
+      editor = parameter.finalEditor,
+      defaultValue = parameter.finalDefaultValue,
       additionalVariables = parameter.additionalVariables.mapValuesNow(_.typingResult),
       variablesToHide = parameter.variablesToHide,
       branchParam = parameter.branchParam,