+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/compile/nodecompilation/DynamicNodeValidator.scala
@@ -186,7 +186,7 @@ class DynamicNodeValidator(
       } else {
         val (singleParam, extraNodeParamOpt) = nodeParameters.find(_.name == parameter.name).map((_, None)).getOrElse {
           val paramToAdd =
-            NodeParameter(parameter.name, parameter.defaultValue.getOrElse(Expression.spel("")))
+            NodeParameter(parameter.name, parameter.finalDefaultValue)
           (paramToAdd, Some(paramToAdd))
         }
         val ctxToUse = inputContext match {